Fix row movement and implement comprehensive distance validation system
- Fix row movement functionality to work correctly with single row moves
- Implement comprehensive distance validation with multiple error display
- Add ID-based row identification to handle post-move data consistency
- Refactor HandsontableSchedule component into modular structure:
  - HandsontableConfig.ts: Column configuration and validation
  - HandsontableStyles.ts: CSS style definitions
  - HandsontableUtils.ts: Data transformation and event handlers
- Add debounced distance checking (200ms) for performance optimization
- Implement real-time error highlighting for distance rule violations
- Update specifications to reflect current implementation
- Remove debug console.log statements for production readiness
