+++ CLAUDE.md
@@ -1,0 +1,10 @@
+# Claude Code Settings
+
+## Development Commands
+- Build: `pnpm build`
+- Type Check: `npx tsc -b`
+- Lint: `pnpm lint`
+- Dev Server: `pnpm dev`
+
+## Project Settings
+- Auto-confirmation: Enabled (no "Do you want to proceed?" prompts)
