+++ SPECIFICATIONS.md
@@ -24,11 +24,26 @@
 - **行移動時**: 全ての行の間隔を即座に再計算
 - **行追加/削除時**: 影響を受ける行の間隔を再計算
 
+#### 2.3 データ検証とエラー表示システム
 
-#### 2.3 データ検証とエラー表示
-- **エラー条件**: 下の行の距離が上の行の距離より小さい場合
-- **エラー表示**: 該当行全体の背景色を赤色 (`#fee2e2`) に変更
-- **エラー検知**: 行移動や距離変更の度に自動チェック
+##### 2.3.1 エラー検出条件
+- **距離逆転エラー**: 下の行の距離が上の行の距離より小さい場合
+
+##### 2.3.2 エラー表示仕様
+- **表示方法**: 該当行全体の背景色を赤色 (`#fee2e2`) に変更
+- **複数エラー対応**: 複数行にエラーがある場合、すべての該当行を同時に赤色表示
+- **エラー優先度**: エラー表示は区分別色分けより優先される
+
+##### 2.3.3 エラーチェック実行タイミング
+- **行移動後**: 全行を対象とした包括的チェック
+- **距離編集後**: デバウンス処理（200ms）による最適化チェック
+  - 連続した距離編集の最後の変更から200ms後に実行
+  - パフォーマンス最適化とReact状態更新との競合回避
+
+##### 2.3.4 IDベースの行識別システム
+- **行移動対応**: Handsontableの表示順序とReact状態の不整合を解決
+- **正確な更新**: IDを使用して正しいチェックポイントデータを特定・更新
+- **データ整合性**: 行移動後も正確なエラーチェックと更新を保証
 
 ### 3. 時間計算システム
 
@@ -57,7 +72,15 @@
 ### 5. ユーザーインターフェース仕様
 
 #### 5.1 グリッド操作
-- **行移動**: 最初の2行（集合・スタート）は移動不可
+
+##### 5.1.1 行移動機能
+- **移動方法**: 行番号をドラッグ&ドロップで移動
+- **移動制限**: 最初の2行（集合・スタート）は移動不可
+- **移動後処理**: 
+  - 全行の距離エラーチェックを自動実行
+  - 間隔の再計算は行わない（Handsontableの表示のみ変更）
+
+##### 5.1.2 その他の操作
 - **列移動**: No列（最初の列）は移動不可
 - **列幅調整**: 列境界のドラッグで調整可能
 - **右クリックメニュー**: 行の追加・削除機能
@@ -94,8 +117,17 @@
 - Tailwind CSS (スタイリング)
 
 ### 主要ファイル構成
+
+#### コンポーネント
 - `RunningSchedulePlanner.tsx`: メインコンポーネント
-- `HandsontableSchedule.tsx`: グリッドコンポーネント
+- `HandsontableSchedule.tsx`: グリッドコンポーネント（リファクタリング済み）
+
+#### テーブル関連モジュール（分離済み）
+- `table/HandsontableConfig.ts`: カラム設定とバリデーション
+- `table/HandsontableStyles.ts`: CSS スタイル定義
+- `table/HandsontableUtils.ts`: データ変換とイベントハンドラー
+
+#### その他
 - `useTimeCalculations.ts`: 時間計算フック
 - `ConfigManager.ts`: プリセット管理サービス
 
@@ -115,7 +147,16 @@ pnpm build
 ```
 
 ## 制約事項
-- 最初の2行（集合・スタート）は削除・移動不可
-- 距離は前の行以上の値である必要がある
-- 間隔列は自動計算のため直接編集不可
-- 計算結果の列（日付、到着、出発）は読み取り専用
+
+### データ制約
+- **距離順序**: 距離は前の行以上の値である必要がある
+- **保護行**: 最初の2行（集合・スタート）は削除・移動不可
+- **読み取り専用列**: 間隔列、計算結果列（日付、到着、出発）は直接編集不可
+
+### パフォーマンス制約
+- **デバウンス処理**: 距離編集後のエラーチェックは200ms遅延
+- **全行チェック**: 行移動時は全行を対象とした処理実行
+
+### 技術制約
+- **IDベース管理**: チェックポイントの識別と更新はID基準
+- **状態同期**: Handsontableの表示状態とReact状態の分離管理
